fix: move supervisor into metastore (#2674)
closes https://github.com/GlareDB/glaredb/issues/2649
