+++ crates/catalog/Cargo.toml
@@ -18,5 +18,3 @@ tonic = { workspace = true }
 tracing = { workspace = true }
 uuid = { version = "1.7.0", features = ["v4", "fast-rng", "macro-diagnostics"] }
 
-[dev-dependencies]
-metastore = { path = "../metastore" }
