+++ crates/metastore/Cargo.toml
@@ -29,6 +29,7 @@ once_cell = "1.19.0"
 proptest-derive = "0.4"
 tower = "0.4"
 dashmap = "5.5.0"
+catalog = { path = "../catalog" }
 
 [dev-dependencies]
 proptest = "1.4"
