+++ crates/metastore/src/lib.rs
@@ -1,8 +1,8 @@
 //! The metastore crate defines the service for managing database catalogs.
+pub mod client;
+mod database;
 pub mod errors;
 pub mod local;
 pub mod srv;
-pub mod util;
-
-mod database;
 mod storage;
+pub mod util;
